% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BriDGE.R
\docType{package}
\name{BriDGE-package}
\alias{BriDGE}
\alias{BriDGE-package}
\title{BriDGE: Behavioural research by integrating DAGs and GAMs in Experiments}
\description{
A comprehensive package for causal analysis of randomized controlled trial data.
}
\details{
It features causal discovery, mediation analysis with nonlinear relationships using GAMs,
bootstrapping, sensitivity analysis, and visualization tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gav888/BriDGE}
  \item Report bugs at \url{https://github.com/gav888/BriDGE/issues}
}

}
\author{
\strong{Maintainer}: Giuseppe Veltri \email{gaveltri@nus.edu.sg} (\href{https://orcid.org/0000-0002-9472-2236}{ORCID})

Authors:
\itemize{
  \item Giuseppe Veltri \email{gaveltri@nus.edu.sg} (\href{https://orcid.org/0000-0002-9472-2236}{ORCID})
}

}
\keyword{internal}
