% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BriDGE.R
\name{bridge_compare}
\alias{bridge_compare}
\title{Comparative Analysis}
\usage{
bridge_compare(data, treatment, mediators, outcome)
}
\arguments{
\item{data}{A data frame containing the variables}

\item{treatment}{Character string specifying the treatment variable name}

\item{mediators}{Character vector specifying the mediator variable names}

\item{outcome}{Character string specifying the outcome variable name}
}
\value{
A data frame containing comparative statistics
}
\description{
Performs comparative analysis between treatment and control groups.
}
