% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_server.R
\name{dep_heatmap_server}
\alias{dep_heatmap_server}
\title{Server module for DEP heatmap + volcano
Expects:
  rv$tables[[tbl]], rv$time_cols[[tbl]], rv$datatype[[tbl]]
  rv$ht_matrix[[tbl]], rv$optimal_k_individual[[tbl]]}
\usage{
dep_heatmap_server(input, output, session, rv, cache, db_path)
}
\description{
Server module for DEP heatmap + volcano
Expects:
  rv$tables[[tbl]], rv$time_cols[[tbl]], rv$datatype[[tbl]]
  rv$ht_matrix[[tbl]], rv$optimal_k_individual[[tbl]]
}
