% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_BRIDGE_INTCLUST}
\alias{reference_BRIDGE_INTCLUST}
\title{Reference matrix for BRIDGE-INTCLUST}
\format{
A numeric matrix with genes in rows and subtype features in columns.
}
\usage{
data(reference_BRIDGE_INTCLUST)
}
\description{
Gene expression reference signatures used for BRIDGE deconvolution (IntClust).
}
\examples{
data(reference_BRIDGE_INTCLUST)
dim(reference_BRIDGE_INTCLUST)
}
\keyword{datasets}
