% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_BRIDGE_PAM50}
\alias{reference_BRIDGE_PAM50}
\title{Reference matrix for BRIDGE-PAM50}
\format{
A numeric matrix with genes in rows and subtype features in columns.
}
\usage{
data(reference_BRIDGE_PAM50)
}
\description{
Gene expression reference signatures used for BRIDGE deconvolution (PAM50).
}
\examples{
data(reference_BRIDGE_PAM50)
dim(reference_BRIDGE_PAM50)
}
\keyword{datasets}
