% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_BRIDGE_TNBC}
\alias{reference_BRIDGE_TNBC}
\title{Reference matrix for BRIDGE-TNBC}
\format{
A numeric matrix with genes in rows and subtype features in columns.
}
\usage{
data(reference_BRIDGE_TNBC)
}
\description{
Gene expression reference signatures used for BRIDGE deconvolution (TNBC).
}
\examples{
data(reference_BRIDGE_TNBC)
dim(reference_BRIDGE_TNBC)
}
\keyword{datasets}
